+++ vsintegration/src/FSharp.Editor/LanguageService/WorkspaceExtensions.fs
@@ -116,7 +116,7 @@ type Document with
                     let! ct = Async.CancellationToken
                     match! projectOptionsManager.TryGetOptionsForDocumentOrProject(this, ct, userOpName) with
                     | None -> return raise(System.OperationCanceledException("FSharp project options not found."))
-                    | Some(parsingOptions, _, projectOptions) ->
+                    | Some(parsingOptions, projectOptions) ->
                         let result = (service.Checker, projectOptionsManager, parsingOptions, projectOptions)
                         return ProjectCache.Projects.GetValue(this.Project, Runtime.CompilerServices.ConditionalWeakTable<_,_>.CreateValueCallback(fun _ -> result))
             else
@@ -127,7 +127,7 @@ type Document with
     member this.GetFSharpCompilationDefinesAsync(userOpName) =
         async {
             let! _, _, parsingOptions, _ = this.GetFSharpCompilationOptionsAsync(userOpName)
-            return CompilerEnvironment.GetCompilationDefinesForEditing parsingOptions
+            return CompilerEnvironment.GetConditionalDefinesForEditing parsingOptions
         }
 
     /// Get the instance of the FSharpChecker from the workspace by the given F# document.
@@ -206,7 +206,7 @@ type Document with
     /// This is only used for testing purposes. It sets the ProjectCache.Projects with the given FSharpProjectOptions and F# document's project.
     member this.SetFSharpProjectOptionsForTesting(projectOptions: FSharpProjectOptions) =
         let workspaceService = this.Project.Solution.GetFSharpWorkspaceService()
-        let parsingOptions, _, _ = 
+        let parsingOptions, _ = 
             workspaceService.FSharpProjectOptionsManager.TryGetOptionsForDocumentOrProject(this, CancellationToken.None, nameof(this.SetFSharpProjectOptionsForTesting))
             |> Async.RunImmediateExceptOnUI
             |> Option.get